alloy_zksync::contracts::common::erc20::ERC20

Enum ERC20Calls

pub enum ERC20Calls {
    allowance(allowanceCall),
    approve(approveCall),
    name(nameCall),
    symbol(symbolCall),
    decimals(decimalsCall),
}

Container for all the ERC20 function calls.

Variants

allowance(allowanceCall)

approve(approveCall)

name(nameCall)

symbol(symbolCall)

decimals(decimalsCall)

Implementations

impl ERC20Calls

pub const SELECTORS: &'static [[u8; 4]]

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.