alloy_zksync

Module provider

ZKsync provider.

Modules

fillers

ZKsync-specific transaction data filler.

layers

ZKsync-specific provider layers.

Structs

DepositExecutor

Type that handles deposit logic for various scenarios: deposit ETH, ERC20 etc.

DepositRequest

Type for deposit request. This type only stores the required information for the deposit, while the deposit itself is performed via DepositExecutor.

L1TransactionReceipt

A wrapper struct to hold L1 transaction receipt and L2 provider which is used by the associated functions.

Enums

L1CommunicationError

Enum to describe errors that might occur during L1 -> L2 communication.

Traits

ProviderBuilderExt

ZKsync-specific extensions for the ProviderBuilder.

ZksyncProvider

ZKsync provider is an extension trait for the Provider trait that adds ZKsync-specific methods.

ZksyncProviderWithWallet

Trait for ZKsync provider with populated wallet Contains provider methods that need a wallet

Functions

zksync_provider

Convenience function to initialize provider builder for ZKsync network.