alloy_zksync::provider::layers::anvil_zksync

Struct AnvilZKsyncProvider

pub struct AnvilZKsyncProvider<P> { /* private fields */ }

A provider that wraps an AnvilZKsyncInstance, preventing the instance from being dropped while the provider is in use.

Implementations

impl<P> AnvilZKsyncProvider<P>
where P: Provider<Zksync>,

pub fn new(inner: P, _anvil: Arc<AnvilZKsyncInstance>) -> Self

Creates a new AnvilZKsyncProvider with the given inner provider and anvil instance.