Struct alloy_zksync::provider::layers::era_test_node::EraTestNodeProvider

pub struct EraTestNodeProvider<P, T> { /* private fields */ }

A provider that wraps an EraTestNodeInstance, preventing the instance from being dropped while the provider is in use.

Implementations

impl<P, T> EraTestNodeProvider<P, T>
where P: Provider<T, Zksync>, T: Transport + Clone,

pub fn new(inner: P, _anvil: Arc<EraTestNodeInstance>) -> Self

Creates a new EraTestNodeProvider with the given inner provider and anvil instance.