alloy_zksync::provider

Trait ZksyncProviderWithWallet

pub trait ZksyncProviderWithWallet<T = BoxTransport>: ZksyncProvider + WalletProvider<Zksync> {
    // Provided method
    fn deposit<'life0, 'life1, 'life2, 'async_trait, P>(
        &'life0 self,
        deposit_request: &'life1 DepositRequest,
        l1_provider: &'life2 P,
    ) -> Pin<Box<dyn Future<Output = Result<L1TransactionReceipt, L1CommunicationError>> + Send + 'async_trait>>
       where P: Provider<Ethereum> + 'async_trait,
             Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
}

Trait for ZKsync provider with populated wallet Contains provider methods that need a wallet

Provided Methods

fn deposit<'life0, 'life1, 'life2, 'async_trait, P>( &'life0 self, deposit_request: &'life1 DepositRequest, l1_provider: &'life2 P, ) -> Pin<Box<dyn Future<Output = Result<L1TransactionReceipt, L1CommunicationError>> + Send + 'async_trait>>
where P: Provider<Ethereum> + 'async_trait, Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Deposits specified L1 token to the L2 address.

Parameters

deposit_request: deposit request which contains deposit params including amount, token to deposit etc.

l1_provider: reference to the L1 provider.

Returns

L1TransactionReceipt. Hint: use returned L1 transaction receipt to get corresponding L2 transaction and wait for its receipt E.g.: deposit_l1_receipt.get_l2_tx()?.with_required_confirmations(1).with_timeout(Some(std::time::Duration::from_secs(60 * 5))).get_receipt()

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.