alloy_zksync::contracts::common::erc20

Module ERC20

ABI for an ERC20 contract.

contract ERC20 {
    function allowance(address owner, address spender) external view returns (uint256);
    function approve(address spender, uint256 value) external returns (bool);
    function name() public view virtual returns (string memory);
    function symbol() public view virtual returns (string memory);
    function decimals() public view virtual returns (uint8);
}

Structs

ERC20Instance

A ERC20 instance.

allowanceCall

Function with signature allowance(address,address) and selector 0xdd62ed3e.

allowanceReturn

Container type for the return parameters of the allowance(address,address) function.

approveCall

Function with signature approve(address,uint256) and selector 0x095ea7b3.

approveReturn

Container type for the return parameters of the approve(address,uint256) function.

decimalsCall

Function with signature decimals() and selector 0x313ce567.

decimalsReturn

Container type for the return parameters of the decimals() function.

nameCall

Function with signature name() and selector 0x06fdde03.

nameReturn

Container type for the return parameters of the name() function.

symbolCall

Function with signature symbol() and selector 0x95d89b41.

symbolReturn

Container type for the return parameters of the symbol() function.

Enums

ERC20Calls

Container for all the ERC20 function calls.

Functions

new

Creates a new wrapper around an on-chain ERC20 contract instance.